cloudrun/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
firebase-admin==7.1.0
# google-genai==1.62.0


annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2025.10.0
google-auth==2.48.0
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
rsa==4.9.1
soupsieve==2.8.3
sympy==1.14.0
tenacity==9.1.2
torch==2.8.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
